Viewport kinda works with new mvi editor architecture

+++ SynapseEngine/EditorCore/Types/TextureHandle.h
@@ -2,6 +2,6 @@
 #include <cstdint>
 
 namespace Syn {
-    using TextureHandle = uint32_t;
+    using TextureHandle = uint64_t;
     constexpr TextureHandle InvalidTextureHandle = 0;
 }
